\name{FLUSH-package}
\alias{FLUSH-package}
\alias{FLUSH}
\docType{package}
\title{Package for Filtering out Likely Uninformative Set of Hybridizations

}
\description{
This package implements functions and graphics for filtering features
in Affymetrix array experiments, based on information content at the
probe level.
}
\details{
\tabular{ll}{
Package: \tab FLUSH\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2007-06-28\cr
License: \tab GPL2 or newer\cr
}
}
\author{
Stefano Calza, Yudi Pawitan

Maintainer: Stefano Calza  <calza@med.unibs.it>

}
\references{
Stefano Calza, Wolfgang Raffelsberger, Alexander Ploner, Jose Sahel, Thierry Leveillard
and Yudi Pawitan. Filtering genes to improve sensitivity in oligonucleotide
microarray data analysis - 2007 Submitted
}
\keyword{ package }
\seealso{
}
\examples{
}
